Decompose provisionLab into phase modules over an ARM toolkit
The 1,889-line provisionLab function becomes a ~150-line sequencer
over ten phase modules (foundation/storage/networking/monitoring/
analytics/flowlogs/compute/dcr/cribl/gateway-phase.ts) behind the
LabPhaseContext seam. The repeated shapes get one home each in
arm-resource.ts: ensureResource (GET-first idempotent ensure, strict
404 and lenient any-non-2xx miss modes) and pollProvisioningState
(attempt-bounded, seeded/unseeded, keep-vs-reset on failed reads).
Public types move to provision-lab-types.ts and are re-exported from
provision-lab.ts unchanged.

All 25 engine tests pass unchanged as characterization pins; the new
arm-resource.test.ts pins the toolkit seam directly (13 tests).
